expose useRouterPush as asyncRouter hook, and improve job table on row click
