+++ web-app/src/app/app/agents/[agentId]/jobs/@right/components/job-input/util.ts
@@ -1,6 +1,3 @@
-import { useRouter } from "next/navigation";
-import { useEffect, useState, useTransition } from "react";
-
 import {
   JobInputOptionSchemaType,
   JobInputSchemaType,
@@ -32,73 +29,3 @@ export const isSingleOption = (
       validation === ValidJobInputValidationTypes.MAX && value <= 1,
   );
 };
-
-export function useRouterRefresh() {
-  const router = useRouter();
-  const [isPending, startTransition] = useTransition();
-
-  const [resolve, setResolve] = useState<((value: unknown) => void) | null>(
-    null,
-  );
-  const [isTriggered, setIsTriggered] = useState(false);
-
-  const refresh = () => {
-    return new Promise((resolve, _reject) => {
-      setResolve(() => resolve);
-      startTransition(() => {
-        router.refresh();
-      });
-    });
-  };
-
-  useEffect(() => {
-    if (isTriggered && !isPending) {
-      if (resolve) {
-        resolve(null);
-
-        setIsTriggered(false);
-        setResolve(null);
-      }
-    }
-    if (isPending) {
-      setIsTriggered(true);
-    }
-  }, [isTriggered, isPending, resolve]);
-
-  return refresh;
-}
-
-export function useRouterPush() {
-  const router = useRouter();
-  const [isPending, startTransition] = useTransition();
-
-  const [resolve, setResolve] = useState<((value: unknown) => void) | null>(
-    null,
-  );
-  const [isTriggered, setIsTriggered] = useState(false);
-
-  const push = (path: string) => {
-    return new Promise((resolve, _reject) => {
-      setResolve(() => resolve);
-      startTransition(() => {
-        router.push(path);
-      });
-    });
-  };
-
-  useEffect(() => {
-    if (isTriggered && !isPending) {
-      if (resolve) {
-        resolve(null);
-
-        setIsTriggered(false);
-        setResolve(null);
-      }
-    }
-    if (isPending) {
-      setIsTriggered(true);
-    }
-  }, [isTriggered, isPending, resolve]);
-
-  return push;
-}
